+++ android/src/main/java/com/ahm/capacitor/camera/preview/camera1api/Camera1Activity.java
@@ -130,11 +130,11 @@ public class Camera1Activity extends Fragment {
         return view;
     }
 
-    public void setRect(int x, int y, int width, int height) {
+    public void setRect(int x, int y, Integer width, Integer height) {
         this.x = x;
         this.y = y;
-        this.width = width;
-        this.height = height;
+        this.width = width == null ? ViewGroup.LayoutParams.MATCH_PARENT : width;
+        this.height = height == null ? ViewGroup.LayoutParams.MATCH_PARENT : height;
     }
 
     private void createCameraPreview() {
@@ -167,8 +167,7 @@ public class Camera1Activity extends Fragment {
     private void setupTouchAndBackButton() {
         final GestureDetector gestureDetector = new GestureDetector(getActivity().getApplicationContext(), new TapGestureDetector());
 
-        getActivity()
-            .runOnUiThread(
+        getActivity().runOnUiThread(
                 new Runnable() {
                     @Override
                     public void run() {