+++ android/src/main/java/com/ahm/capacitor/camera/preview/camera1api/Camera1Preview.java
@@ -609,36 +609,21 @@ public class Camera1Preview implements Camera1Activity.CameraPreviewListener {
                         int computedY = (int) TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_DIP, y, metrics);
 
                         // size
-                        int computedWidth;
-                        int computedHeight;
-                        int computedPaddingBottom;
+                        Integer computedWidth = null;
+                        Integer computedHeight = null;
+                        int computedPaddingBottom = 0;
 
                         if (paddingBottom != 0) {
                             computedPaddingBottom = (int) TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_DIP, paddingBottom, metrics);
-                        } else {
-                            computedPaddingBottom = 0;
                         }
 
                         if (width != 0) {
                             computedWidth = (int) TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_DIP, width, metrics);
-                        } else {
-                            Display defaultDisplay = plugin.getBridge().getActivity().getWindowManager().getDefaultDisplay();
-                            final Point size = new Point();
-                            defaultDisplay.getSize(size);
-
-                            computedWidth = (int) TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_PX, size.x, metrics);
                         }
 
                         if (height != 0) {
                             computedHeight =
                                 (int) TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_DIP, height, metrics) - computedPaddingBottom;
-                        } else {
-                            Display defaultDisplay = plugin.getBridge().getActivity().getWindowManager().getDefaultDisplay();
-                            final Point size = new Point();
-                            defaultDisplay.getSize(size);
-
-                            computedHeight =
-                                (int) TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_PX, size.y, metrics) - computedPaddingBottom;
                         }
 
                         fragment.setRect(computedX, computedY, computedWidth, computedHeight);