+++ ios/Tests/CameraPreviewPluginTests/CameraPreviewPluginTests.swift
@@ -1,0 +1,5 @@
+import XCTest
+@testable import CameraPreviewPlugin
+
+class CameraPreviewTests: XCTestCase {
+}